Move Logout button to Header
