+++ client/src/Header.jsx
@@ -5,12 +5,21 @@ import { connect } from 'react-redux';
 import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
+import useReactRouter from 'use-react-router';
 
+import { Button } from '@material-ui/core';
 import NavMenu from './NavMenu';
 
-function Header({ loggedIn, name }) {
+import { logoutThunk } from './actionCreators';
+
+function Header({ loggedIn, name, dispatchLogoutThunk }) {
+  const { history } = useReactRouter();
+  function onLogout() {
+    dispatchLogoutThunk();
+    history.push('/');
+  }
   return (
-    <div>
+    <div className="app-bar">
       <AppBar position="static">
         <Toolbar disableGutters>
           <Typography variant="h6" color="inherit" style={{ flex: 1 }}>
@@ -19,6 +28,11 @@ function Header({ loggedIn, name }) {
           <Typography variant="h6" color="inherit">
             {loggedIn ? name : ''}
           </Typography>
+          {loggedIn && (
+            <Button variant="contained" onClick={onLogout}>
+              Logout
+            </Button>
+          )}
         </Toolbar>
       </AppBar>
       <NavMenu />
@@ -34,14 +48,20 @@ const mapStateToProps = ({ profile }) => ({
 Header.propTypes = {
   loggedIn: PropTypes.bool,
   name: PropTypes.string,
+  dispatchLogoutThunk: PropTypes.func,
 };
 
 Header.defaultProps = {
   loggedIn: false,
   name: '',
+  dispatchLogoutThunk: () => {},
+};
+
+const mapDispatchToProps = {
+  dispatchLogoutThunk: logoutThunk,
 };
 
 export default connect(
   mapStateToProps,
-  null,
+  mapDispatchToProps,
 )(Header);
